+++ Runtime/!Core/Types/Singletons/SingletonObject.cs
@@ -1,0 +1,31 @@
+using Sirenix.OdinInspector;
+using UnityEngine;
+
+namespace Shadowpaw {
+  /// <summary>
+  /// A Singleton ScriptableObject
+  /// </summary>
+  /// <remarks>
+  /// This is a base class for creating Singleton ScriptableObjects in Unity.
+  /// Instances are lazily created or loaded from the Resources folder when accessed.
+  /// </remarks>
+  public abstract class SingletonObject<T> : RootObject where T : SingletonObject<T> {
+    /// <summary>
+    /// A Singleton Instance
+    /// </summary>
+    public static T Instance => Singletons.GetOrCreate(() => {
+      // Search for an existing instance
+      var resources = Resources.LoadAll<T>("");
+      if (resources.Length > 1) {
+        Debug.LogWarning($"Multiple instances of Singleton {typeof(T).Name} found in Resources folder.");
+      }
+      if (resources.Length > 0) return resources[0];
+
+      // Unity gets grumpy about creating generic instances, so we can't do that...
+      if (typeof(T).IsGenericType) return null;
+
+      // Create a new non-generic instance
+      return CreateInstance<T>();
+    });
+  }
+}