+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Initial release of Shadowpaw - Unity Toolkit
 - `Interfaces`: Added interfaces for ILocator, ILocator{T}, and ISubtypeLocator{TBase}
 - `Interfaces`: Added interfaces for IProvider, IProvider{T}, and ISubtypeProvider{TBase}
 - `Types`: Added Proxy types for Unity's MonoBehaviour and ScriptableObject
+- `Types`: Added Singleton types for Behaviours and Objects
 - `Utils`: Added extensions for LayerMask, List, Reflection, String, Task, and Type
 
 Be good to each other 💕